Updatd Hooks1.cs
